requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
cryptography==2.7
ExifRead==2.1.2
Fabric3==1.14.post1
Flask==1.0.2
Flask-Uploads==0.2.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paramiko==2.4.2
passlib==1.7.1
Pillow==5.3.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
Werkzeug==0.14.1
WTForms==2.2.1
